Added a small effect on the search criteria items.
The effect is located at #Search_Sidebar_Criteria selectors;
The darker border fades in slowly and fades out even slower.
A bad thing that is happening is that the edges are totally squared
when fading out, I could not fix that. We need to figure out how to
make the edges invisible.