Experimenting on positioning:fixed for our Constant View Banner.
We can easily have it fixed on top of the page at all times (even if one scroll the page).

But then we would need to set the rest to be absolute instead of static, which means we have to set their x and y positions, as well as their width and length (e.g. for the content view).