Styled the Search results Table.
Now it looks fancier. I made some logic to alternate the background color
for every row and created a shared set of action buttons (show, edit,
remove) instead of having a set of buttons at every row.
This required me to add some logic for allowing selection of a result.
The selected entry is a little more opace than others, to differentiate.