Changed review controller code that was broken, incorrect, or didn't make any sense.
There were some syntax errors in the controller.  Also, there were views for create, update, and destroy, which are not actions which require views, so I removed them.

There is a lot to be done to get reviews to go with listing, etc.. but that can wait for another day.